sydbox: assorted improvements

+++ src/pathdecode.c
@@ -16,7 +16,7 @@
  * -errno : Negated errno indicating error code
  *  0     : Successful run
  */
-int path_decode(syd_process_t *current, unsigned arg_index, char **buf)
+int path_decode(syd_process_t *current, uint8_t arg_index, char **buf)
 {
 	int r;
 	ssize_t count_read;
@@ -59,7 +59,7 @@ int path_decode(syd_process_t *current, unsigned arg_index, char **buf)
  * -errno : Negated errno indicating error code
  *  0     : Successful run
  */
-int path_prefix(syd_process_t *current, unsigned arg_index, char **buf)
+int path_prefix(syd_process_t *current, uint8_t arg_index, char **buf)
 {
 	int r, fd;
 	char *prefix = NULL;
